split rigid body runtime file
