+++ src/wasm_src/src/runtime/mod.rs
@@ -4,4 +4,6 @@ pub(crate) mod multi_physics_runtime;
 pub(crate) mod multi_physics_world;
 pub(crate) mod physics_runtime;
 pub(crate) mod physics_world;
+pub(crate) mod rigidbody_bundle;
+pub(crate) mod rigidbody_construction_info;
 pub(crate) mod rigidbody;
